Split prepare_grid into its five windowed passes
430 lines and six nesting levels in one function. Each pass — release, evict,
build, the focused card's tiles, the shared ones, the reveal — becomes a method
in render/prepare_grid.rs; the window arithmetic stays in prepare_grid alone.
No logic change: art_ready becomes a free fn so both users can call it while
&mut self is live, and the reveal now runs after the shared tiles instead of
between them (neither depends on the other).
