+++ src/app/render/prepare.rs
@@ -8,7 +8,6 @@ use std::time::Instant;
 
 use anyhow::Result;
 
-use crate::app::grid::{CARD_BUILD_BUDGET, CARD_BUILD_BURST, CARD_KEEP_ROWS, CARD_PREFETCH_ROWS};
 use crate::app::hosts::HostEntry;
 use crate::app::nav::ScreenKey;
 use crate::app::render::ctx::RenderCtx;
@@ -96,435 +95,6 @@ impl App {
                 )
             })? {
                 updated.push(tile::FOCUS_ROW);
-            }
-        }
-        Ok(())
-    }
-
-    /// Whether this frame's grid pass can be skipped outright: a modal owns the screen, the
-    /// grid has already revealed, and nothing has invalidated a card behind it.
-    fn grid_window_frozen(&self) -> bool {
-        !matches!(self.nav.screen, Screen::Home)
-            && self.render.grid.reveal.is_revealed()
-            && !self.render.grid.dirty
-            && self.render.grid.cards_dirty.is_empty()
-    }
-
-    /// Grid family: windowed/budgeted card-tile building, eviction, the reveal
-    /// spinner, and the shared ring/outline/pin-badge tiles — or the "no host"
-    /// hint when nothing is selected. Extracted from `prepare_tiles` (A2 staging).
-    fn prepare_grid(&mut self, ctx: &mut RenderCtx<'_>) -> Result<()> {
-        let RenderCtx {
-            tiles,
-            text: text_cache,
-            fonts,
-            screen: size,
-            updated,
-            ..
-        } = ctx;
-        let (screen_w, screen_h) = (size.w, size.h);
-        // The same three numbers `advance_frame` sized `self.render.grid.card_size` from — the grid's
-        // whole geometry follows from the width it has to fill.
-        let available_w = screen_w.saturating_sub(ui::widgets::SIDEBAR_W);
-        let columns = view::home::grid_columns(available_w);
-        let (card_w, card_h) = view::home::grid_card_size(available_w, columns);
-        // Reset before the branch: it is only ever set inside it, and a stale `true` left
-        // behind by a host that has since been deselected would spin the render loop at
-        // full rate forever.
-        self.render.grid.tiles_pending = false;
-        if self.library.selected_host.is_some() {
-            // Nothing behind an open modal can come into view: the grid neither scrolls nor
-            // moves focus while a modal owns input, so the whole windowed pass — the one cost
-            // here that scales with the window — is skipped unless a card was actually
-            // invalidated. The grid still composites under the modal's scrim from the tiles it
-            // already holds.
-            if self.grid_window_frozen() {
-                return Ok(());
-            }
-            let count = self.grid_len(columns);
-            // Captured before it's cleared below: a fresh library load is the only
-            // rebuild that also re-arms the spinner.
-            let full_reset = self.render.grid.dirty;
-            if full_reset {
-                // Every existing texture is stale (different games, different host) —
-                // drop them rather than leaving them to be overwritten one by one,
-                // which would strand the tail of a longer previous library.
-                for id in self.render.grid.card_ids.release_all() {
-                    tiles.remove(id);
-                    self.render.evicted_tiles.push(id);
-                }
-                self.render.grid.card_pop_until = None;
-                self.render.grid.dirty = false;
-                self.render.grid.cards_dirty.clear();
-                // Scrolling or re-pinning a card must not hide the already-visible
-                // grid behind the spinner again.
-                self.render.grid.reveal.restart();
-            } else {
-                for id in std::mem::take(&mut self.render.grid.cards_dirty) {
-                    if let Some(t) = self.render.grid.card_ids.release(&id) {
-                        tiles.remove(t);
-                    }
-                }
-            }
-
-            // Windowed, budgeted tile building — see `CARD_BUILD_BUDGET`. Both windows are
-            // index ranges, so every pass below iterates the window rather than the library.
-            let row_h = card_h as i32 + view::home::GRID_GAP;
-            let visible_rows = (screen_h as i32 - view::home::GRID_TOP_Y).max(row_h) / row_h + 1;
-            let first_visible_row = (self.render.grid.scroll / row_h).max(0);
-            let rows = count.div_ceil(columns.max(1)) as i32;
-            // Row band -> index range, clamped to the library. Deliberately ignores the
-            // section headings' offsets: a row's worth of slack either way is what
-            // `CARD_PREFETCH_ROWS`/`CARD_KEEP_ROWS` already exist to absorb.
-            let window = |lo: i32, hi: i32| {
-                let lo = lo.clamp(0, rows) as usize * columns.max(1);
-                let hi = (hi + 1).clamp(0, rows) as usize * columns.max(1);
-                lo.min(count)..hi.min(count)
-            };
-            // `mut` because the reveal check below consumes it as an iterator.
-            let mut build_window = window(
-                first_visible_row - CARD_PREFETCH_ROWS,
-                first_visible_row + visible_rows + CARD_PREFETCH_ROWS,
-            );
-            let keep_window = window(
-                first_visible_row - CARD_KEEP_ROWS,
-                first_visible_row + visible_rows + CARD_KEEP_ROWS,
-            );
-
-            // Held by value, not re-derived per index — and, unlike the `App`
-            // helpers, it maps indices without borrowing all of `self`, so the art
-            // lookups below can sit next to `&mut self.jobs.art`.
-            let layout = self.grid_layout(columns);
-
-            // Evict first, so a long scroll frees textures in the same frame it needs new
-            // ones rather than a frame later.
-            //
-            // Driven off what is resident (`CardIds`) against the keep window, not off a scan
-            // of the whole library: the resident set is bounded by the window, so this costs
-            // the window twice over however large the library gets.
-            //
-            // By tile id rather than by pin id: the ids are `Copy` integers, so the keep set
-            // is a sorted window-sized vector and the test is a binary search — where a
-            // `HashSet<&str>` re-hashed every resident card's id string every frame. Both
-            // lists are `GridState` scratch, cleared and refilled rather than allocated.
-            let mut keep = std::mem::take(&mut self.render.grid.scratch.keep);
-            keep.clear();
-            keep.extend(
-                keep_window
-                    .filter_map(|idx| layout.pin_id_at(&self.library.games, idx))
-                    .filter_map(|id| self.render.grid.card_ids.get(id)),
-            );
-            keep.sort_unstable();
-            let mut dropped = std::mem::take(&mut self.render.grid.scratch.dropped);
-            dropped.clear();
-            dropped.extend(
-                self.render.grid
-                    .card_ids
-                    .entries()
-                    .filter(|(_, t)| keep.binary_search(t).is_err())
-                    .map(|(id, _)| id.to_string()),
-            );
-            self.render.grid.scratch.keep = keep;
-            for id in dropped.drain(..) {
-                if let Some(t) = self.render.grid.card_ids.release(&id) {
-                    // Kept for the cards built below: a scroll frees and needs the same
-                    // card-sized pixmap in the same frame (see `GridState::free_cards`).
-                    if let Some(painter) = tiles.take(t) {
-                        self.render.grid.free_cards.push(painter);
-                    }
-                    self.render.evicted_tiles.push(t);
-                }
-                // Drop the decoded cover too — it is several times the size of the tile it
-                // feeds. Re-requested from the disk cache on scroll back. (Nothing to drop for
-                // the pinned "Desktop" entry, which has no art at all.)
-                self.library.art.remove(&id);
-                if let Some(loader) = &mut self.jobs.art {
-                    loader.forget(&id);
-                }
-            }
-
-            // Ready once nothing more can arrive: cover already in `self.library.art`, or the game
-            // never had one to fetch (no `self.library.art` entry either way). "Desktop" and the
-            // padding after a partial pinned row have no `games` entry and are always ready.
-            let art_ready = |idx: usize| {
-                layout.game_at(&self.library.games, idx).is_none_or(|game| {
-                    self.library.art.contains_key(&game.id) || (game.art.portrait.is_none() && game.art.header.is_none())
-                })
-            };
-
-            // Art-ready cards build first — building one before its cover arrives just
-            // burns a second budget slot re-dirtying it once the cover shows up. Two lists
-            // rather than one sorted one, and indices rather than ids: a candidate past the
-            // budget is never built, so nothing should be copied on its behalf.
-            let mut ready = std::mem::take(&mut self.render.grid.scratch.ready);
-            let mut waiting = std::mem::take(&mut self.render.grid.scratch.waiting);
-            ready.clear();
-            waiting.clear();
-            for idx in build_window.clone() {
-                // Nothing to build or fetch art for in the padding after a partial
-                // pinned row.
-                let Some(id) = layout.pin_id_at(&self.library.games, idx) else {
-                    continue;
-                };
-                // Ask for this card's cover as it enters the window, not for the whole
-                // library at once (see `art::ArtLoader`).
-                if let (Some(loader), Some(game)) = (&mut self.jobs.art, layout.game_at(&self.library.games, idx)) {
-                    loader.request(game);
-                }
-                if self.render.grid.card_ids.get(id).is_some_and(|t| tiles.contains(t)) {
-                    continue;
-                }
-                if art_ready(idx) {
-                    ready.push(idx);
-                } else {
-                    waiting.push(idx);
-                }
-            }
-
-            let mut pending = false;
-            let budget_from = Instant::now();
-            let mut built = 0usize;
-            for idx in ready.iter().copied().chain(waiting.iter().copied()) {
-                // Counted on cards actually rasterized, not on candidates seen: the budget is
-                // a time budget, and skipping a padding slot costs none of it.
-                if built >= CARD_BUILD_BURST || (built > 0 && budget_from.elapsed() >= CARD_BUILD_BUDGET) {
-                    pending = true;
-                    break;
-                }
-                let Some(id) = layout.pin_id_at(&self.library.games, idx).map(str::to_string) else {
-                    continue;
-                };
-                built += 1;
-                let recycled = self.render.grid.free_cards.pop();
-                let tile = {
-                    let (title, art) = self.grid_card_content(idx, columns);
-                    ui::rasterize_into(
-                        ui::tiles::CardTile {
-                            w: card_w,
-                            h: card_h,
-                            title,
-                            art,
-                        },
-                        recycled,
-                        text_cache,
-                        fonts,
-                    )?
-                };
-                let tile_id = self.render.grid.card_ids.id(&id);
-                tiles.put(tile_id, cache::STATIC, tile);
-                if self.render.grid.reveal.is_revealed() {
-                    self.render.grid.arm_card_pop(&id, Instant::now());
-                }
-                updated.push(tile_id);
-            }
-            // Anything still here evicted without a card being built in its place, so it is
-            // surplus rather than churn — held past the frame it would just be a cache of
-            // pixmaps nothing asked for.
-            self.render.grid.free_cards.clear();
-            self.render.grid.scratch.dropped = dropped;
-            self.render.grid.scratch.ready = ready;
-            self.render.grid.scratch.waiting = waiting;
-            self.render.grid.tiles_pending = pending;
-
-            // Prefetch the focused card's hero, so the connecting screen has one ready the
-            // moment OK is pressed. Deduped in the loader, and the fetched bytes are
-            // disk-cached, so hovering back over a card costs no round trip.
-            //
-            // Only once the visible window has settled: the loader serves hero requests
-            // ahead of card art, so queueing one mid-scroll would put the cards the user is
-            // actually looking at behind a full-screen fetch and decode.
-            if self.render.grid.reveal.is_revealed() && !pending {
-                if let HomeFocus::Grid(focus_idx) = self.home_focus {
-                    if let Some(game) = layout.game_at(&self.library.games, focus_idx) {
-                        if let Some(loader) = &mut self.jobs.art {
-                            loader.request_hero(game);
-                        }
-                        self.render.hero.want(&game.id);
-                    }
-                }
-            }
-
-            // The focused card's title strip: its own tile, so the wipe in `draw_list` is
-            // a moving source/destination rect — one small blur per focus move instead of
-            // re-rasterizing the card every animation frame.
-            if let HomeFocus::Grid(idx) = self.home_focus {
-                if let Some(pin_id) = layout.pin_id_at(&self.library.games, idx) {
-                    let (title, art) = self.grid_card_content(idx, columns);
-                    // Keyed by card identity like the card tiles themselves (`CardIds`),
-                    // not by title — two games can share one.
-                    let overridden = self.game_has_overrides(pin_id);
-                    let version = cache::version(&(pin_id, card_w, card_h, art.is_some(), overridden));
-                    if tiles.ensure(tile::CARD_TITLE, version, || {
-                        ui::rasterize(
-                            ui::tiles::CardTitleTile {
-                                card_w,
-                                card_h,
-                                title,
-                                art,
-                                overridden,
-                            },
-                            text_cache,
-                            fonts,
-                        )
-                    })? {
-                        updated.push(tile::CARD_TITLE);
-                    }
-
-                    // The submenu panel a hold raises: the same strip grown to carry the
-                    // rows (see `ui::tiles::CardMenuTile`), on the same wipe.
-                    //
-                    // Built when the card takes focus, not when the menu opens — it costs a
-                    // full-card art rescale plus a radius-6 blur, and paying that on the
-                    // frame the rise starts is what made the panel appear to wait for the
-                    // button to come back up. Held off until the grid has settled so a fast
-                    // scroll doesn't pay it per card, unless a menu is already up (which
-                    // can only happen on a settled grid anyway).
-                    let menu_open = self.card_menu.as_ref().is_some_and(|m| m.pin_id == pin_id);
-                    if menu_open || (self.render.grid.reveal.is_revealed() && !pending) {
-                        let rows = self.card_menu_rows(pin_id);
-                        // No focused row in the key: the selection is a `DrawCmd::Fill` laid
-                        // over this tile, so moving between the menu's rows rebuilds nothing.
-                        let version = cache::version(&(pin_id, card_w, card_h, art.is_some(), &rows, overridden));
-                        if tiles.ensure(tile::CARD_MENU, version, || {
-                            ui::rasterize(
-                                ui::tiles::CardMenuTile {
-                                    card_w,
-                                    card_h,
-                                    title,
-                                    art,
-                                    rows: &rows,
-                                },
-                                text_cache,
-                                fonts,
-                            )
-                        })? {
-                            updated.push(tile::CARD_MENU);
-                        }
-                        if tiles.ensure(tile::CARD_MENU_ROWS, version, || {
-                            ui::rasterize(
-                                ui::tiles::CardMenuRowsTile {
-                                    card_w,
-                                    card_h,
-                                    rows: &rows,
-                                    // The dot follows what owns it: the title while the strip is
-                                    // collapsed, the Settings row once the panel is up.
-                                    marked: overridden.then_some(crate::app::state::cardmenu::ROW_SETTINGS),
-                                },
-                                text_cache,
-                                fonts,
-                            )
-                        })? {
-                            updated.push(tile::CARD_MENU_ROWS);
-                        }
-                        if tiles.ensure(tile::CARD_MENU_TITLE, version, || {
-                            ui::rasterize(
-                                ui::tiles::CardMenuTitleTile { card_w, card_h, title },
-                                text_cache,
-                                fonts,
-                            )
-                        })? {
-                            updated.push(tile::CARD_MENU_TITLE);
-                        }
-                        // The selection band's rounded-bottom variant, for the row that ends
-                        // on the card's bottom edge. Keyed by width alone: it is one flat
-                        // colour, so nothing else about the card changes it.
-                        if tiles.ensure(tile::CARD_MENU_BAND, cache::version(&card_w), || {
-                            ui::rasterize(ui::tiles::CardMenuBandTile { card_w }, text_cache, fonts)
-                        })? {
-                            updated.push(tile::CARD_MENU_BAND);
-                        }
-                    }
-                }
-            }
-
-            // The grid's section headings. Built unconditionally like the pin badge: two
-            // lines of text, and whether they are *drawn* is the compose path's call.
-            for (id, label) in [
-                (tile::SECTION_PINNED, crate::app::view::home::SECTION_PINNED_LABEL),
-                (tile::SECTION_LIBRARY, crate::app::view::home::SECTION_LIBRARY_LABEL),
-            ] {
-                if tiles.ensure_static(id, || {
-                    ui::rasterize(
-                        ui::tiles::TextTile {
-                            font: fonts.title,
-                            text: label,
-                            color: ui::style::theme().muted,
-                        },
-                        text_cache,
-                        fonts,
-                    )
-                })? {
-                    updated.push(id);
-                }
-            }
-
-            // The pinned badge tile — built once, composited over the focused
-            // card in `draw_list` rather than baked into individual card tiles.
-            if tiles.ensure_static(tile::PIN_BADGE, || {
-                ui::rasterize(ui::tiles::PinBadgeTile, text_cache, fonts)
-            })? {
-                updated.push(tile::PIN_BADGE);
-            }
-
-            if !self.render.grid.reveal.is_revealed() {
-                // Rechecks the whole window rather than trusting `!pending`, since a card built
-                // earlier can still be waiting behind a re-dirtied sibling; requires `art_ready`
-                // too so a placeholder built this tick can't count as revealed.
-                let window_ready = || {
-                    build_window.all(|idx| {
-                        layout.pin_id_at(&self.library.games, idx).is_none_or(|id| {
-                            self.render.grid.card_ids.get(id).is_some_and(|t| tiles.contains(t)) && art_ready(idx)
-                        })
-                    })
-                };
-                let next_frame = self.render.grid.reveal.advance(self.library_fetch_in_flight(), window_ready);
-                match next_frame {
-                    Some(idx) => updated.push(tile::spinner(idx)),
-                    // Everything built behind the spinner becomes visible in this one frame, so
-                    // it all zooms in off a single clock.
-                    None if self.render.grid.reveal.is_revealed() => {
-                        let now = Instant::now();
-                        let ids: Vec<String> = self.render.grid.card_ids.pin_ids().map(str::to_string).collect();
-                        for id in ids {
-                            self.render.grid.arm_card_pop_if_idle(&id, now);
-                        }
-                    }
-                    None => {}
-                }
-            }
-
-            // One shared tile at the current card size, so the card size *is* the
-            // version — a resolution change rebuilds it, nothing else does.
-            let size = cache::version(&(card_w, card_h));
-            if tiles.ensure(tile::RING, size, || {
-                ui::rasterize(ui::tiles::FocusRingTile { w: card_w, h: card_h }, text_cache, fonts)
-            })? {
-                updated.push(tile::RING);
-            }
-            if tiles.ensure(tile::CARD_SHADOW, size, || {
-                ui::rasterize(ui::tiles::CardShadowTile { w: card_w, h: card_h }, text_cache, fonts)
-            })? {
-                updated.push(tile::CARD_SHADOW);
-            }
-            if tiles.ensure(tile::CARD_OUTLINE, size, || {
-                ui::rasterize(ui::tiles::CardOutlineTile { w: card_w, h: card_h }, text_cache, fonts)
-            })? {
-                updated.push(tile::CARD_OUTLINE);
-            }
-        } else {
-            self.render.grid.reveal.reveal();
-            if tiles.ensure_static(tile::NO_HOST, || {
-                ui::rasterize(
-                    ui::tiles::TextTile {
-                        font: fonts.label,
-                        text: "No host selected — pick one from the list, or add one.",
-                        color: ui::style::theme().muted,
-                    },
-                    text_cache,
-                    fonts,
-                )
-            })? {
-                updated.push(tile::NO_HOST);
             }
         }
         Ok(())
